requirements.txt
mss
keyboard
google-genai
openai
requests
anthropic
Pillow
customtkinter
